+++ Backend/Data/receipt.go
@@ -27,7 +27,7 @@ el monto total a cobrar por la reparacion
 package data
 
 type Receipt struct {
-	Id               uint32
+	Id               int64
 	OperationsAmount int32
 	RechargeAmount   int32
 	DiscountAmount   int32
