+++ Backend/Internal/Tariffs/TariffMileage/tariff_mileage_service.go
@@ -8,11 +8,11 @@ type TariffMileage = data.TariffMileage
 
 type ITariffMileageRepository interface {
 	SaveTariffMileage(TariffMileage) (*TariffMileage, error)
-	GetTariffMileageById(uint32) (*TariffMileage, error)
+	GetTariffMileageById(int64) (*TariffMileage, error)
 	GetTariffMileageCell(string, int32) (*TariffMileage, error)
 	GetAllTariffMileage() ([]TariffMileage, error)
 	UpdateTariffMileage(TariffMileage) error
-	DeleteTariffMileageById(uint32) error
+	DeleteTariffMileageById(int64) error
 }
 
 type Service struct {
@@ -23,7 +23,7 @@ func (serv *Service) SaveTariffMileage(t TariffMileage) (*TariffMileage, error) 
 	return serv.Repository.SaveTariffMileage(t)
 }
 
-func (serv *Service) GetTariffMileageById(id uint32) (*TariffMileage, error) {
+func (serv *Service) GetTariffMileageById(id int64) (*TariffMileage, error) {
 	return serv.Repository.GetTariffMileageById(id)
 }
 
@@ -43,6 +43,6 @@ func (serv *Service) UpdateTariffMileage(t TariffMileage) error {
 	return serv.Repository.UpdateTariffMileage(t)
 }
 
-func (serv *Service) DeleteTariffMileageById(id uint32) error {
+func (serv *Service) DeleteTariffMileageById(id int64) error {
 	return serv.Repository.DeleteTariffMileageById(id)
 }
